+++ angular/projects/catalyst/src/lib/dialog/dialog.component.ts
@@ -25,6 +25,13 @@ import { CatDialogHeaderComponent } from './dialog-header.component';
   }
 })
 export class CatDialogComponent implements AfterContentInit {
+
+  /**
+   * Flag to disable scrolling behavior of dialog content. Default is false. If set to true the
+   * using component is responsible for not overflowing the viewport.
+   */
+  @Input() noScroll = false;
+
   /**
    * Flag to disable/enable overscroll behavior.
    */