+++ core/src/components.d.ts
@@ -339,7 +339,7 @@ export namespace Components {
          */
         "doFocus": (options?: FocusOptions) => Promise<void>;
         /**
-          * Fine-grained control over when the errors are shown. Can be `false` to never show errors, `true` to show errors on blur, or a number to show errors on change with the given delay in milliseconds.
+          * Fine-grained control over when the errors are shown. Can be `false` to never show errors, `true` to show errors on blur, or a number to show errors change with the given delay in milliseconds or immediately on blur.
          */
         "errorUpdate": boolean | number;
         /**
@@ -529,7 +529,7 @@ export namespace Components {
          */
         "doFocus": (options?: FocusOptions) => Promise<void>;
         /**
-          * Fine-grained control over when the errors are shown. Can be `false` to never show errors, `true` to show errors on blur, or a number to show errors on change with the given delay in milliseconds.
+          * Fine-grained control over when the errors are shown. Can be `false` to never show errors, `true` to show errors on blur, or a number to show errors change with the given delay in milliseconds or immediately on blur.
          */
         "errorUpdate": boolean | number;
         /**
@@ -777,7 +777,7 @@ export namespace Components {
          */
         "doFocus": (options?: FocusOptions) => Promise<void>;
         /**
-          * Fine-grained control over when the errors are shown. Can be `false` to never show errors, `true` to show errors on blur, or a number to show errors on change with the given delay in milliseconds.
+          * Fine-grained control over when the errors are shown. Can be `false` to never show errors, `true` to show errors on blur, or a number to show errors change with the given delay in milliseconds or immediately on blur.
          */
         "errorUpdate": boolean | number;
         /**
@@ -1093,7 +1093,7 @@ export namespace Components {
          */
         "doFocus": (options?: FocusOptions) => Promise<void>;
         /**
-          * Fine-grained control over when the errors are shown. Can be `false` to never show errors, `true` to show errors on blur, or a number to show errors on change with the given delay in milliseconds.
+          * Fine-grained control over when the errors are shown. Can be `false` to never show errors, `true` to show errors on blur, or a number to show errors change with the given delay in milliseconds or immediately on blur.
          */
         "errorUpdate": boolean | number;
         /**
@@ -1266,6 +1266,9 @@ export namespace Components {
          */
         "tabsAlign": 'left' | 'center' | 'right' | 'justify';
     }
+    /**
+     * An input that allows multiple values to be entered as tags.
+     */
     interface CatTag {
         /**
           * Whether the input should show a clear button.
@@ -1276,7 +1279,7 @@ export namespace Components {
          */
         "disabled": boolean;
         /**
-          * Fine-grained control over when the errors are shown. Can be `false` to never show errors, `true` to show errors on blur, or a number to show errors on change with the given delay in milliseconds.
+          * Fine-grained control over when the errors are shown. Can be `false` to never show errors, `true` to show errors on blur, or a number to show errors change with the given delay in milliseconds or immediately on blur.
          */
         "errorUpdate": boolean | number;
         /**
@@ -1352,7 +1355,7 @@ export namespace Components {
          */
         "doFocus": (options?: FocusOptions) => Promise<void>;
         /**
-          * Fine-grained control over when the errors are shown. Can be `false` to never show errors, `true` to show errors on blur, or a number to show errors on change with the given delay in milliseconds.
+          * Fine-grained control over when the errors are shown. Can be `false` to never show errors, `true` to show errors on blur, or a number to show errors change with the given delay in milliseconds or immediately on blur.
          */
         "errorUpdate": boolean | number;
         /**
@@ -1450,7 +1453,7 @@ export namespace Components {
          */
         "doFocus": (options?: FocusOptions) => Promise<void>;
         /**
-          * Fine-grained control over when the errors are shown. Can be `false` to never show errors, `true` to show errors on blur, or a number to show errors on change with the given delay in milliseconds.
+          * Fine-grained control over when the errors are shown. Can be `false` to never show errors, `true` to show errors on blur, or a number to show errors change with the given delay in milliseconds or immediately on blur.
          */
         "errorUpdate": boolean | number;
         /**
@@ -2171,6 +2174,9 @@ declare global {
         "catFocus": FocusEvent;
         "catBlur": FocusEvent;
     }
+    /**
+     * An input that allows multiple values to be entered as tags.
+     */
     interface HTMLCatTagElement extends Components.CatTag, HTMLStencilElement {
         addEventListener<K extends keyof HTMLCatTagElementEventMap>(type: K, listener: (this: HTMLCatTagElement, ev: CatTagCustomEvent<HTMLCatTagElementEventMap[K]>) => any, options?: boolean | AddEventListenerOptions): void;
         addEventListener<K extends keyof DocumentEventMap>(type: K, listener: (this: Document, ev: DocumentEventMap[K]) => any, options?: boolean | AddEventListenerOptions): void;
@@ -2605,7 +2611,7 @@ declare namespace LocalJSX {
          */
         "disabled"?: boolean;
         /**
-          * Fine-grained control over when the errors are shown. Can be `false` to never show errors, `true` to show errors on blur, or a number to show errors on change with the given delay in milliseconds.
+          * Fine-grained control over when the errors are shown. Can be `false` to never show errors, `true` to show errors on blur, or a number to show errors change with the given delay in milliseconds or immediately on blur.
          */
         "errorUpdate"?: boolean | number;
         /**
@@ -2784,7 +2790,7 @@ declare namespace LocalJSX {
          */
         "disabled"?: boolean;
         /**
-          * Fine-grained control over when the errors are shown. Can be `false` to never show errors, `true` to show errors on blur, or a number to show errors on change with the given delay in milliseconds.
+          * Fine-grained control over when the errors are shown. Can be `false` to never show errors, `true` to show errors on blur, or a number to show errors change with the given delay in milliseconds or immediately on blur.
          */
         "errorUpdate"?: boolean | number;
         /**
@@ -3031,7 +3037,7 @@ declare namespace LocalJSX {
          */
         "disabled"?: boolean;
         /**
-          * Fine-grained control over when the errors are shown. Can be `false` to never show errors, `true` to show errors on blur, or a number to show errors on change with the given delay in milliseconds.
+          * Fine-grained control over when the errors are shown. Can be `false` to never show errors, `true` to show errors on blur, or a number to show errors change with the given delay in milliseconds or immediately on blur.
          */
         "errorUpdate"?: boolean | number;
         /**
@@ -3371,7 +3377,7 @@ declare namespace LocalJSX {
          */
         "disabled"?: boolean;
         /**
-          * Fine-grained control over when the errors are shown. Can be `false` to never show errors, `true` to show errors on blur, or a number to show errors on change with the given delay in milliseconds.
+          * Fine-grained control over when the errors are shown. Can be `false` to never show errors, `true` to show errors on blur, or a number to show errors change with the given delay in milliseconds or immediately on blur.
          */
         "errorUpdate"?: boolean | number;
         /**
@@ -3558,6 +3564,9 @@ declare namespace LocalJSX {
          */
         "tabsAlign"?: 'left' | 'center' | 'right' | 'justify';
     }
+    /**
+     * An input that allows multiple values to be entered as tags.
+     */
     interface CatTag {
         /**
           * Whether the input should show a clear button.
@@ -3568,7 +3577,7 @@ declare namespace LocalJSX {
          */
         "disabled"?: boolean;
         /**
-          * Fine-grained control over when the errors are shown. Can be `false` to never show errors, `true` to show errors on blur, or a number to show errors on change with the given delay in milliseconds.
+          * Fine-grained control over when the errors are shown. Can be `false` to never show errors, `true` to show errors on blur, or a number to show errors change with the given delay in milliseconds or immediately on blur.
          */
         "errorUpdate"?: boolean | number;
         /**
@@ -3643,7 +3652,7 @@ declare namespace LocalJSX {
          */
         "disabled"?: boolean;
         /**
-          * Fine-grained control over when the errors are shown. Can be `false` to never show errors, `true` to show errors on blur, or a number to show errors on change with the given delay in milliseconds.
+          * Fine-grained control over when the errors are shown. Can be `false` to never show errors, `true` to show errors on blur, or a number to show errors change with the given delay in milliseconds or immediately on blur.
          */
         "errorUpdate"?: boolean | number;
         /**
@@ -3740,7 +3749,7 @@ declare namespace LocalJSX {
          */
         "disabled"?: boolean;
         /**
-          * Fine-grained control over when the errors are shown. Can be `false` to never show errors, `true` to show errors on blur, or a number to show errors on change with the given delay in milliseconds.
+          * Fine-grained control over when the errors are shown. Can be `false` to never show errors, `true` to show errors on blur, or a number to show errors change with the given delay in milliseconds or immediately on blur.
          */
         "errorUpdate"?: boolean | number;
         /**
@@ -4099,6 +4108,9 @@ declare module "@stencil/core" {
              * window, using tabs as a navigational element.
              */
             "cat-tabs": LocalJSX.CatTabs & JSXBase.HTMLAttributes<HTMLCatTabsElement>;
+            /**
+             * An input that allows multiple values to be entered as tags.
+             */
             "cat-tag": LocalJSX.CatTag & JSXBase.HTMLAttributes<HTMLCatTagElement>;
             /**
              * Textarea specifies a control that allows user to write text over multiple