+++ core/src/components/cat-input/cat-input.tsx
@@ -32,7 +32,7 @@ export class CatInput {
   }
 
   private input!: HTMLInputElement;
-  private errorMapSrc?: ErrorMap;
+  private errorMapSrc?: ErrorMap | true;
 
   @Element() hostElement!: HTMLElement;
 
@@ -42,7 +42,7 @@ export class CatInput {
 
   @State() isPasswordShown = false;
 
-  @State() errorMap?: ErrorMap;
+  @State() errorMap?: ErrorMap | true;
 
   /**
    * Whether the label need a marker to shown if the input is required or optional.
@@ -187,7 +187,7 @@ export class CatInput {
   /**
    * Fine-grained control over when the errors are shown. Can be `false` to
    * never show errors, `true` to show errors on blur, or a number to show
-   * errors on change with the given delay in milliseconds.
+   * errors change with the given delay in milliseconds or immediately on blur.
    */
   @Prop() errorUpdate: boolean | number = 0;
 
@@ -211,8 +211,11 @@ export class CatInput {
    */
   @Event() catBlur!: EventEmitter<FocusEvent>;
 
+  componentWillLoad(): void {
+    this.onErrorsChanged(this.errors, undefined, false);
+  }
+
   componentWillRender(): void {
-    this.onErrorsChanged(this.errors);
     this.hasSlottedLabel = !!this.hostElement.querySelector('[slot="label"]');
     this.hasSlottedHint = !!this.hostElement.querySelector('[slot="hint"]');
   }
@@ -262,16 +265,16 @@ export class CatInput {
   }
 
   @Watch('errors')
-  onErrorsChanged(value?: boolean | string[] | ErrorMap) {
+  onErrorsChanged(newValue?: boolean | string[] | ErrorMap, _oldValue?: unknown, update: boolean = true) {
     if (!coerceBoolean(this.errorUpdate)) {
       this.errorMap = undefined;
     } else {
-      this.errorMapSrc = Array.isArray(value)
-        ? (value as string[]).reduce((acc, err) => ({ ...acc, [err]: undefined }), {})
-        : value === true
-          ? {}
-          : value || undefined;
-      this.showErrorsIfTimeout() || this.showErrorsIfNoFocus();
+      this.errorMapSrc = Array.isArray(newValue)
+        ? (newValue as string[]).reduce((acc, err) => ({ ...acc, [err]: undefined }), {})
+        : newValue || undefined;
+      if (update) {
+        this.showErrorsIfTimeout() || this.showErrorsIfNoFocus();
+      }
     }
   }
 
@@ -414,7 +417,7 @@ export class CatInput {
   }
 
   private get invalid() {
-    return !!Object.keys(this.errorMap || {}).length;
+    return this.errorMap === true || !!Object.keys(this.errorMap || {}).length;
   }
 
   private onInput() {